requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
Augmentor==0.2.3
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
Bottleneck==1.2.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dataclasses==0.6
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.9
entrypoints==0.3
fastai==1.0.52
fastprogress==0.1.21
future==0.17.1
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.20
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
msgpack==0.6.1
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numexpr==2.6.9
numpy==1.16.4
olefile==0.46
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
plac==0.9.6
preshed==2.0.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.14.11
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
pyzmq==18.0.0
qtconsole==4.5.1
regex==2018.1.10
requests==2.22.0
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.8
spacy==2.0.18
terminado==0.8.2
testpath==0.4.2
thinc==6.12.1
toolz==0.9.0
torch==1.1.0
torchfile==0.1.0
torchvision==0.3.0
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
typing==3.6.4
ujson==1.35
urllib3==1.24.2
visdom==0.1.8.8
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
widgetsnbextension==3.4.2
wrapt==1.10.11
